Features added: PLOT takes multiple setup structures and returns processed plot structures.
Multiple setup objects means you have more control over the way that your
functions are sampled and a bunch of other stuff.  Setup structures are used in
the order they appear and apply to everything after them in the argument list.
This is a stop-gap on the path to something better, though.

Returning the plot objects makes it easy to replot later (computed data is saved
in the object).